moved AwsSlice to reducers folder and updated imports
